backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
google-genai>=1.0.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
